Text Processing Functions
Part A: Utilities
Write a method that reads in a text file and returns a list of the
tokens (preferably alphanumeric) in that file. Write a method to print
out frequency results.

Part B: Word Frequencies
Count the total number of words and their frequencies in a token list.

Part C: 2-grams
A 2-gram is two words that occur consecutively in a file. For example,
"two words", "words that", and "that occur" are all 2-grams from the
previous sentence.

Part D: Palindromes
A palindrome is a words or phrase that reads. the same in both
directions. For example, these are all palindromes: "kayak", "Do geese
see god", "A man, a plan, a canal--Panama". Count the total number of
palindromes and their frequencies in a text file.